purchase, purchase item, inventory, and post operations are completed

+++ Learning.DataAccess/category.cs
@@ -6,27 +6,52 @@ using System.Threading.Tasks;
 
 namespace Learning.DataAccess {
   public class category
-    : parent {
-    public void CreateCategories() {
-      Model.Entities.category Beverages = new Model.Entities.category() { name = "Beverages" };
-      _context.Categories.Add(Beverages);
-      _context.SaveChanges();
+    : Base {
+
+    public Model.Entities.category select(int category_id) {
+      Model.Entities.category category =
+        _context.Categories.Find(new object[] { category_id });
+
+      return category;
     }
-
-    public IEnumerable<object> select(Model.Entities.category category) {
-      var categories = _context.Categories.Select(c => new { c.category_id, c.name, c.description }).Where(p => p.category_id == category.category_id || p.name == category.name);
-
-      return categories;
+    public bool insert(Model.Entities.category category) {
+      bool result = false;
+      try {
+        _context.Categories.Add(category);
+        _context.SaveChanges();
+        result = true;
+      } catch (System.Data.Entity.Infrastructure.DbUpdateException) {
+        // log; cannot insert the category
+      }
+      return result;
     }
-
-    public void delect(Model.Entities.category category) {
-      _context.Categories.Remove(category);
-      _context.SaveChanges();
+    public bool update(Model.Entities.category category) {
+      bool result = false;
+      Model.Entities.category temp = select(category.category_id);
+      if (temp != null) {
+        try {
+          temp.name = category.name;
+          temp.description = category.description;
+          _context.SaveChanges();
+        } catch (System.Data.Entity.Infrastructure.DbUpdateException) {
+          // log; error occur while updating the category;
+        }
+      } else {
+        // log; cannot find the category for updating
+      }
+      return result;
     }
-
-    public void insert(Model.Entities.category category) {
-      _context.Categories.Add(category);
-      _context.SaveChanges();
+    public bool delect(int category_id) {
+      bool result = false;
+      Model.Entities.category category = select(category_id);
+      if (category != null) {
+        _context.Categories.Remove(category);
+        _context.SaveChanges();
+        result = true;
+      } else {
+        // log; cannot find the category
+      }
+      return result;
     }
 
   }
