+++ Learning.DataAccess/post.cs
@@ -4,30 +4,60 @@ using System.Linq;
 
 namespace Learning.DataAccess {
   public class post
-    : parent {
+    : Base {
 
     public Model.Entities.post select(int post_type_id, int identifier) {
       Model.Entities.post post = _context.Posts.SingleOrDefault(p => (p.post_type_id == post_type_id & p.identifier == identifier));
       return post;
     }
-    public void insert(Model.Entities.post post) {
-      _context.Posts.Add(post);
-      _context.SaveChanges();
+    public bool insert(Model.Entities.post post) {
+      bool result = false;
+      try {
+        _context.Posts.Add(post);
+        _context.SaveChanges();
+        result = true;
+      } catch (System.Data.Entity.Infrastructure.DbUpdateException) {
+        // log; cannot insert purchase
+
+      }
+      return result;
     }
     public void update(Model.Entities.post post) {
       if (post != null) {
         _context.SaveChanges();
       }
     }
-    public void update(int post_type_id, int identifier, decimal total) {
-      Model.Entities.post temp = select(post_type_id, identifier);
-      temp.crebit = total;
-      _context.SaveChanges();
-    }
-    public void delete(int post_type_id, int identifier) {
+    public bool update(int post_type_id, int identifier, decimal total) {
+      bool result = false;
       Model.Entities.post post = select(post_type_id, identifier);
-      _context.Posts.Remove(post);
-      _context.SaveChanges();
+      if (post != null) {
+        try {
+          post.crebit = total;
+          _context.SaveChanges();
+          result = true;
+        } catch (System.Data.Entity.Infrastructure.DbUpdateException) {
+          // log; cannot update the purchase in post
+        }
+      } else {
+        // log; cannot find the required purchase post
+      }
+      return result;
+    }
+    public bool delete(int post_type_id, int identifier) {
+      bool result = false;
+      Model.Entities.post post = select(post_type_id, identifier);
+      if (post != null) {
+        try {
+          _context.Posts.Remove(post);
+          _context.SaveChanges();
+          result = true;
+        } catch (System.Data.Entity.Infrastructure.DbUpdateException) {
+          // log; cannot update post
+        }
+      } else {
+        // log; cannot find the post to delete
+      }
+      return result;
     }
     public void delete(Model.Entities.post post) {
       if (post != null) {
